Rename tables to better convey their use
The `s_table` is now the `saved_value_table`; it stores the values of
the token strings when they match certain elements of the grammar (like
die size, or modifiers) so they can be used by the `Dice()` class to
construct the dice.

The `p_table` is now the `stack_action_table`; it stores the actions
taken based on the elements on the stack and stream, and pushes more
items onto the stack.